+++ mta/processing.h
@@ -9,7 +9,7 @@
 
 int getHeaders(unsigned char * const data, size_t * const lenData, struct emailInfo * const email);
 void moveHeader(unsigned char * const data, size_t * const lenData, const char * const needle, const size_t lenNeedle, unsigned char * const target, uint8_t * const lenTarget, const size_t limit);
-void decodeEncodedWord(unsigned char * const data, uint8_t * const lenData);
+void decodeEncodedWord(unsigned char * const data, size_t * const lenData);
 unsigned char *decodeCte(const int cte, const unsigned char * const src, size_t * const lenSrc);
 void convertToUtf8(unsigned char ** const src, size_t * const lenSrc, const char * const charset);
 char *getCharset(const char *ct);
